Refactor objc2 directory structure a bit
