+++ crates/objc2/src/__macro_helpers/cache.rs
