+++ crates/objc2/src/lib.rs
@@ -206,7 +206,6 @@ macro_rules! __hash_idents {
 
 #[doc(hidden)]
 pub mod __macro_helpers;
-mod cache;
 mod class_type;
 pub mod declare;
 pub mod encode;
