+++ crates/test-ui/ui/add_method_no_bool.stderr
@@ -18,7 +18,7 @@ error[E0277]: the trait bound `bool: Encode` is not satisfied
           and $N others
   = note: required for `unsafe extern "C" fn(&objc2::runtime::Object, objc2::runtime::Sel, bool)` to implement `MethodImplementation`
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -44,7 +44,7 @@ error[E0277]: the trait bound `bool: Encode` is not satisfied
   = note: required for `bool` to implement `EncodeReturn`
   = note: required for `unsafe extern "C" fn(&objc2::runtime::Object, objc2::runtime::Sel) -> bool` to implement `MethodImplementation`
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
