+++ crates/test-ui/ui/declare_add_bad_method.stderr
@@ -17,7 +17,7 @@ error[E0277]: the trait bound `fn(_, _): MethodImplementation` is not satisfied
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -41,7 +41,7 @@ error[E0277]: the trait bound `fn(_, _): MethodImplementation` is not satisfied
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -65,7 +65,7 @@ error[E0277]: the trait bound `fn(_, _) -> _: MethodImplementation` is not satis
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -89,7 +89,7 @@ error[E0277]: the trait bound `fn(_, _) -> _: MethodImplementation` is not satis
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -113,7 +113,7 @@ error[E0277]: the trait bound `fn(_, _, _): MethodImplementation` is not satisfi
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
