+++ crates/test-ui/ui/declare_class_invalid_receiver.stderr
@@ -24,7 +24,7 @@ error[E0277]: the trait bound `extern "C" fn(Box<CustomObject>, objc2::runtime::
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -56,7 +56,7 @@ error[E0277]: the trait bound `extern "C" fn(Id<CustomObject>, objc2::runtime::S
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -88,7 +88,7 @@ error[E0277]: the trait bound `extern "C" fn(CustomObject, objc2::runtime::Sel):
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -120,7 +120,7 @@ error[E0277]: the trait bound `extern "C" fn(Box<CustomObject>, objc2::runtime::
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -152,7 +152,7 @@ error[E0277]: the trait bound `extern "C" fn(Id<CustomObject>, objc2::runtime::S
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
@@ -184,7 +184,7 @@ error[E0277]: the trait bound `extern "C" fn(CustomObject, objc2::runtime::Sel) 
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
