+++ crates/test-ui/ui/extern_methods_invalid_type.stderr
@@ -45,7 +45,7 @@ error[E0277]: the trait bound `i32: MaybeUnwrap` is not satisfied
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -68,7 +68,7 @@ error[E0308]: mismatched types
   = note:           expected enum `Option<Allocated<_>>`
           found mutable reference `&mut MyObject`
 note: associated function defined here
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |               ^^^^^^^^^^^^^^^
@@ -121,7 +121,7 @@ error[E0277]: the trait bound `Result<Id<MyObject>, Id<NSObject>>: MaybeUnwrap` 
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
