+++ crates/test-ui/ui/fn_ptr_reference_method.stderr
@@ -17,7 +17,7 @@ error[E0277]: the trait bound `for<'a> extern "C" fn(_, _, &'a objc2::runtime::O
             extern "C" fn(&'a T, objc2::runtime::Sel, A, B, C, D, E, F, G) -> R
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
- --> $WORKSPACE/crates/objc2/src/declare.rs
+ --> $WORKSPACE/crates/objc2/src/declare/mod.rs
   |
   |         F: MethodImplementation<Callee = T>,
   |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ClassBuilder::add_method`
