+++ crates/test-ui/ui/msg_send_id_invalid_return.stderr
@@ -10,7 +10,7 @@ error[E0277]: the trait bound `&objc2::runtime::Object: MaybeUnwrap` is not sati
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -64,7 +64,7 @@ error[E0277]: the trait bound `&objc2::runtime::Object: MaybeUnwrap` is not sati
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -97,7 +97,7 @@ error[E0271]: type mismatch resolving `<Id<Object> as MaybeUnwrap>::Input == All
   = note: expected struct `Allocated<_>`
              found struct `Id<objc2::runtime::Object>`
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                                   ^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -112,7 +112,7 @@ error[E0271]: type mismatch resolving `<Id<Allocated<Object>> as MaybeUnwrap>::I
   = note: expected struct `Allocated<_>`
              found struct `Id<Allocated<objc2::runtime::Object>>`
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                                   ^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -130,7 +130,7 @@ error[E0277]: the trait bound `&objc2::runtime::Object: MaybeUnwrap` is not sati
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -145,7 +145,7 @@ error[E0271]: type mismatch resolving `<Id<Class> as MaybeUnwrap>::Input == Id<O
   = note: expected struct `Id<objc2::runtime::Object, _>`
              found struct `Id<objc2::runtime::Class, Shared>`
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                                   ^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -160,7 +160,7 @@ error[E0271]: type mismatch resolving `<Id<NSObject> as MaybeUnwrap>::Input == I
   = note: expected struct `Id<objc2::runtime::Object, _>`
              found struct `Id<NSObject, Shared>`
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                                   ^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -178,7 +178,7 @@ error[E0277]: the trait bound `&objc2::runtime::Object: MaybeUnwrap` is not sati
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -196,7 +196,7 @@ error[E0277]: the trait bound `&objc2::runtime::Object: MaybeUnwrap` is not sati
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
@@ -212,7 +212,7 @@ error[E0277]: the trait bound `Option<&objc2::runtime::Object>: MaybeUnwrap` is 
             Option<Allocated<T>>
             Option<Id<T, O>>
 note: required by a bound in `send_message_id`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |     unsafe fn send_message_id<A: MessageArguments, R: MaybeUnwrap<Input = U>>(
   |                                                       ^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `MsgSendId::send_message_id`
