+++ crates/test-ui/ui/msg_send_no_return_type.stderr
@@ -6,7 +6,7 @@ error[E0282]: type annotations needed
   |
   = note: this error originates in the macro `$crate::__msg_send_helper` which comes from the expansion of the macro `msg_send` (in Nightly builds, run with -Z macro-backtrace for more info)
 help: consider giving `result` an explicit type
- --> $WORKSPACE/crates/objc2/src/macros.rs
+ --> $WORKSPACE/crates/objc2/src/macros/mod.rs
   |
   |         let result: /* Type */;
   |                   ++++++++++++
