+++ crates/test-ui/ui/msg_send_underspecified_error.stderr
@@ -36,7 +36,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |         *mut *mut E: Encode,
   |                      ^^^^^^ required by this bound in `MsgSendId::send_message_id_error`
@@ -54,7 +54,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |         *mut *mut E: Encode,
   |                      ^^^^^^ required by this bound in `MsgSendId::send_message_id_error`
@@ -72,7 +72,7 @@ error[E0283]: type annotations needed
             where T: RefEncode, T: ?Sized;
   = note: required for `*mut *mut _` to implement `Encode`
 note: required by a bound in `send_message_id_error`
- --> $WORKSPACE/crates/objc2/src/__macro_helpers.rs
+ --> $WORKSPACE/crates/objc2/src/__macro_helpers/mod.rs
   |
   |         *mut *mut E: Encode,
   |                      ^^^^^^ required by this bound in `MsgSendId::send_message_id_error`
